N-06: fix `getSchedule` parameter type mismatch
Change `getSchedule(uint256 scheduleId)` function parameter type to
`uint32` to match the `Schedule` struct's `id` field type. This prevents
potential implicit downcasting issues.

+++ contracts/vaults/VestingVault.sol
@@ -717,7 +717,7 @@ contract VestingVault is Context, AccessControl, SalvageCapable, IVestingVault, 
      * @param scheduleId The ID of the schedule
      * @return schedule The schedule with the specified ID
      */
-    function getSchedule(uint256 scheduleId) external view override returns (Schedule memory schedule) {
+    function getSchedule(uint32 scheduleId) external view override returns (Schedule memory schedule) {
         schedule = scheduleById[scheduleId];
         if (schedule.id == 0) revert LibErrors.NotFound(scheduleId);
         return schedule;
